backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-multipart==0.0.20
pydantic-settings==2.9.1
email-validator==2.2.0
httpx==0.28.1
PyMuPDF==1.25.5
